requirements.txt
asgiref==3.4.0
dj-database-url==0.5.0
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
future==0.18.2
gunicorn==20.1.0
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
yamlordereddictloader==0.4.0
